move FormElement component to separate file
